cluster-unification: split generic parts of mz_compute::server into the mz_cluster crate
